OF-3101: Admin console shows user vCard / profile info
This adds a page named 'Profile Information' to the admin console (under 'Users/Groups' > 'Users' > 'User Options').

The new page displays vCard information for the associated user.
