+++ i18n/src/main/resources/openfire_i18n.properties
@@ -145,6 +145,8 @@ tab.tab-users.descr=Click to manage users and groups
         sidebar.user-summary=User Summary
         sidebar.user-summary.descr=Click to see a list of users in the system
             sidebar.sidebar-users-options=User Options
+                sidebar.user-profile=Profile Information
+                sidebar.user-profile.descr=Click to view the user&#39;s profile information
                 sidebar.user-properties=User Properties
                 sidebar.user-properties.descr=Click to edit the user&#39;s properties
                 sidebar.user-roster=Roster
@@ -2649,6 +2651,11 @@ user.privacylist.is_default_list=(this is the default list)
 user.privacylists.no_items_on_list=There are no items on this privacy list.
 user.privacylist.allow=ALLOW
 user.privacylist.deny=DENY
+
+# User profile Page
+
+user.profile.title=Profile Information
+user.profile.info=This page shows the profile information that is published for this user. This information is typically available to the user's contacts.
 
 # User properties Page
