+++ xmppserver/src/main/java/org/jivesoftware/admin/LdapUserProfile.java
@@ -20,6 +20,7 @@ import org.dom4j.*;
 import org.dom4j.io.OutputFormat;
 import org.jivesoftware.openfire.ldap.LdapManager;
 import org.jivesoftware.openfire.ldap.LdapVCardProvider;
+import org.jivesoftware.openfire.vcard.VCardBean;
 import org.jivesoftware.openfire.vcard.VCardManager;
 import org.jivesoftware.util.JiveGlobals;
 import org.jivesoftware.util.XMLWriter;
@@ -36,245 +37,11 @@ import java.util.*;
  *
  * @author Gaston Dombiak
  */
-public class LdapUserProfile {
-
+public class LdapUserProfile extends VCardBean
+{
     private static final Logger Log = LoggerFactory.getLogger(LdapUserProfile.class);
 
-    private String name = "";
-    private String email = "";
-    private String fullName = "";
-    private String nickname = "";
-    private String birthday = "";
-    private String photo = "";
-    private String homeStreet = "";
-    private String homeCity = "";
-    private String homeState = "";
-    private String homeZip = "";
-    private String homeCountry = "";
-    private String homePhone = "";
-    private String homeMobile = "";
-    private String homeFax = "";
-    private String homePager = "";
-    private String businessStreet = "";
-    private String businessCity = "";
-    private String businessState = "";
-    private String businessZip = "";
-    private String businessCountry = "";
-    private String businessJobTitle = "";
-    private String businessDepartment = "";
-    private String businessPhone = "";
-    private String businessMobile = "";
-    private String businessFax = "";
-    private String businessPager = "";
     private Boolean avatarStoredInDB = false;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getFullName() {
-        return fullName;
-    }
-
-    public void setFullName(String fullName) {
-        this.fullName = fullName;
-    }
-
-    public String getNickname() {
-        return nickname;
-    }
-
-    public void setNickname(String nickname) {
-        this.nickname = nickname;
-    }
-
-    public String getBirthday() {
-        return birthday;
-    }
-
-    public void setBirthday(String birthday) {
-        this.birthday = birthday;
-    }
-
-    public String getPhoto() {
-        return photo;
-    }
-
-    public void setPhoto(String photo) {
-        this.photo = photo;
-    }
-
-    public String getHomeStreet() {
-        return homeStreet;
-    }
-
-    public void setHomeStreet(String homeStreet) {
-        this.homeStreet = homeStreet;
-    }
-
-    public String getHomeCity() {
-        return homeCity;
-    }
-
-    public void setHomeCity(String homeCity) {
-        this.homeCity = homeCity;
-    }
-
-    public String getHomeState() {
-        return homeState;
-    }
-
-    public void setHomeState(String homeState) {
-        this.homeState = homeState;
-    }
-
-    public String getHomeZip() {
-        return homeZip;
-    }
-
-    public void setHomeZip(String homeZip) {
-        this.homeZip = homeZip;
-    }
-
-    public String getHomeCountry() {
-        return homeCountry;
-    }
-
-    public void setHomeCountry(String homeCountry) {
-        this.homeCountry = homeCountry;
-    }
-
-    public String getHomePhone() {
-        return homePhone;
-    }
-
-    public void setHomePhone(String homePhone) {
-        this.homePhone = homePhone;
-    }
-
-    public String getHomeMobile() {
-        return homeMobile;
-    }
-
-    public void setHomeMobile(String homeMobile) {
-        this.homeMobile = homeMobile;
-    }
-
-    public String getHomeFax() {
-        return homeFax;
-    }
-
-    public void setHomeFax(String homeFax) {
-        this.homeFax = homeFax;
-    }
-
-    public String getHomePager() {
-        return homePager;
-    }
-
-    public void setHomePager(String homePager) {
-        this.homePager = homePager;
-    }
-
-    public String getBusinessStreet() {
-        return businessStreet;
-    }
-
-    public void setBusinessStreet(String businessStreet) {
-        this.businessStreet = businessStreet;
-    }
-
-    public String getBusinessCity() {
-        return businessCity;
-    }
-
-    public void setBusinessCity(String businessCity) {
-        this.businessCity = businessCity;
-    }
-
-    public String getBusinessState() {
-        return businessState;
-    }
-
-    public void setBusinessState(String businessState) {
-        this.businessState = businessState;
-    }
-
-    public String getBusinessZip() {
-        return businessZip;
-    }
-
-    public void setBusinessZip(String businessZip) {
-        this.businessZip = businessZip;
-    }
-
-    public String getBusinessCountry() {
-        return businessCountry;
-    }
-
-    public void setBusinessCountry(String businessCountry) {
-        this.businessCountry = businessCountry;
-    }
-
-    public String getBusinessJobTitle() {
-        return businessJobTitle;
-    }
-
-    public void setBusinessJobTitle(String businessJobTitle) {
-        this.businessJobTitle = businessJobTitle;
-    }
-
-    public String getBusinessDepartment() {
-        return businessDepartment;
-    }
-
-    public void setBusinessDepartment(String businessDepartment) {
-        this.businessDepartment = businessDepartment;
-    }
-
-    public String getBusinessPhone() {
-        return businessPhone;
-    }
-
-    public void setBusinessPhone(String businessPhone) {
-        this.businessPhone = businessPhone;
-    }
-
-    public String getBusinessMobile() {
-        return businessMobile;
-    }
-
-    public void setBusinessMobile(String businessMobile) {
-        this.businessMobile = businessMobile;
-    }
-
-    public String getBusinessFax() {
-        return businessFax;
-    }
-
-    public void setBusinessFax(String businessFax) {
-        this.businessFax = businessFax;
-    }
-
-    public String getBusinessPager() {
-        return businessPager;
-    }
-
-    public void setBusinessPager(String businessPager) {
-        this.businessPager = businessPager;
-    }
 
     public Boolean getAvatarStoredInDB() {
         return avatarStoredInDB;
@@ -359,136 +126,8 @@ public class LdapUserProfile {
      * Saves current configuration as XML/DB properties.
      */
     public void saveProperties() {
-        Element vCard = DocumentHelper.createElement(QName.get("vCard", "vcard-temp"));
-        Element subelement;
+        final Element vCard = asElement();
 
-        // Add name
-        if (name != null && !name.trim().isEmpty()) {
-            subelement = vCard.addElement("N");
-            subelement.addElement("GIVEN").setText(name.trim());
-        }
-        // Add email
-        if (email != null && !email.trim().isEmpty()) {
-            subelement = vCard.addElement("EMAIL");
-            subelement.addElement("INTERNET");
-            subelement.addElement("USERID").setText(email.trim());
-        }
-        // Add Full Name
-        vCard.addElement("FN").setText(fullName.trim());
-        // Add nickname
-        if (nickname != null && !nickname.trim().isEmpty()) {
-            vCard.addElement("NICKNAME").setText(nickname.trim());
-        }
-        // Add birthday
-        if (birthday != null && !birthday.trim().isEmpty()) {
-            vCard.addElement("BDAY").setText(birthday.trim());
-        }
-        // Add photo/avatar
-        if (photo != null && !photo.trim().isEmpty()) {
-            Element element = vCard.addElement("PHOTO");
-            element.addElement("TYPE").setText("image/jpeg");
-            element.addElement("BINVAL").setText(photo.trim());
-        }
-        // Add home address
-        subelement = vCard.addElement("ADR");
-        subelement.addElement("HOME");
-        if (homeStreet != null && !homeStreet.trim().isEmpty()) {
-            subelement.addElement("STREET").setText(homeStreet.trim());
-        }
-        if (homeCity != null && !homeCity.trim().isEmpty()) {
-            subelement.addElement("LOCALITY").setText(homeCity.trim());
-        }
-        if (homeState != null && !homeState.trim().isEmpty()) {
-            subelement.addElement("REGION").setText(homeState.trim());
-        }
-        if (homeZip != null && !homeZip.trim().isEmpty()) {
-            subelement.addElement("PCODE").setText(homeZip.trim());
-        }
-        if (homeCountry != null && !homeCountry.trim().isEmpty()) {
-            subelement.addElement("CTRY").setText(homeCountry.trim());
-        }
-        // Add business address
-        subelement = vCard.addElement("ADR");
-        subelement.addElement("WORK");
-        if (businessStreet != null && !businessStreet.trim().isEmpty()) {
-            subelement.addElement("STREET").setText(businessStreet.trim());
-        }
-        if (businessCity != null && !businessCity.trim().isEmpty()) {
-            subelement.addElement("LOCALITY").setText(businessCity.trim());
-        }
-        if (businessState != null && !businessState.trim().isEmpty()) {
-            subelement.addElement("REGION").setText(businessState.trim());
-        }
-        if (businessZip != null && !businessZip.trim().isEmpty()) {
-            subelement.addElement("PCODE").setText(businessZip.trim());
-        }
-        if (businessCountry != null && !businessCountry.trim().isEmpty()) {
-            subelement.addElement("CTRY").setText(businessCountry.trim());
-        }
-        // Add home phone
-        if (homePhone != null && !homePhone.trim().isEmpty()) {
-            subelement = vCard.addElement("TEL");
-            subelement.addElement("HOME");
-            subelement.addElement("VOICE");
-            subelement.addElement("NUMBER").setText(homePhone.trim());
-        }
-        // Add home mobile
-        if (homeMobile != null && !homeMobile.trim().isEmpty()) {
-            subelement = vCard.addElement("TEL");
-            subelement.addElement("HOME");
-            subelement.addElement("CELL");
-            subelement.addElement("NUMBER").setText(homeMobile.trim());
-        }
-        // Add home fax
-        if (homeFax != null && !homeFax.trim().isEmpty()) {
-            subelement = vCard.addElement("TEL");
-            subelement.addElement("HOME");
-            subelement.addElement("FAX");
-            subelement.addElement("NUMBER").setText(homeFax.trim());
-        }
-        // Add home pager
-        if (homePager != null && !homePager.trim().isEmpty()) {
-            subelement = vCard.addElement("TEL");
-            subelement.addElement("HOME");
-            subelement.addElement("PAGER");
-            subelement.addElement("NUMBER").setText(homePager.trim());
-        }
-        // Add business phone
-        if (businessPhone != null && !businessPhone.trim().isEmpty()) {
-            subelement = vCard.addElement("TEL");
-            subelement.addElement("WORK");
-            subelement.addElement("VOICE");
-            subelement.addElement("NUMBER").setText(businessPhone.trim());
-        }
-        // Add business mobile
-        if (businessMobile != null && !businessMobile.trim().isEmpty()) {
-            subelement = vCard.addElement("TEL");
-            subelement.addElement("WORK");
-            subelement.addElement("CELL");
-            subelement.addElement("NUMBER").setText(businessMobile.trim());
-        }
-        // Add business fax
-        if (businessFax != null && !businessFax.trim().isEmpty()) {
-            subelement = vCard.addElement("TEL");
-            subelement.addElement("WORK");
-            subelement.addElement("FAX");
-            subelement.addElement("NUMBER").setText(businessFax.trim());
-        }
-        // Add business pager
-        if (businessPager != null && !businessPager.trim().isEmpty()) {
-            subelement = vCard.addElement("TEL");
-            subelement.addElement("WORK");
-            subelement.addElement("PAGER");
-            subelement.addElement("NUMBER").setText(businessPager.trim());
-        }
-        // Add job title
-        if (businessJobTitle != null && !businessJobTitle.trim().isEmpty()) {
-            vCard.addElement("TITLE").setText(businessJobTitle.trim());
-        }
-        // Add job department
-        if (businessDepartment != null && !businessDepartment.trim().isEmpty()) {
-            vCard.addElement("ORG").addElement("ORGUNIT").setText(businessDepartment.trim());
-        }
         // Generate content to store in property
         String vcardXML;
         StringWriter writer = new StringWriter();
@@ -539,109 +178,7 @@ public class LdapUserProfile {
             // Parse XML
             Document document = DocumentHelper.parseText(xmlProperty);
             Element vCard = document.getRootElement();
-
-            Element element = vCard.element("N");
-            if (element != null) {
-                name = element.elementTextTrim("GIVEN");
-            }
-            element = vCard.element("EMAIL");
-            if (element != null) {
-                email = element.elementTextTrim("USERID");
-            }
-            element = vCard.element("FN");
-            if (element != null) {
-                fullName = element.getTextTrim();
-            }
-            element = vCard.element("NICKNAME");
-            if (element != null) {
-                nickname = element.getTextTrim();
-            }
-            element = vCard.element("BDAY");
-            if (element != null) {
-                birthday = element.getTextTrim();
-            }
-            // Parse addresses
-            Iterator<Element> addresses = vCard.elementIterator("ADR");
-            while (addresses.hasNext()) {
-                element = addresses.next();
-                if (element.element("HOME") != null) {
-                    if (element.element("STREET") != null) {
-                        homeStreet = element.elementTextTrim("STREET");
-                    }
-                    if (element.element("LOCALITY") != null) {
-                        homeCity = element.elementTextTrim("LOCALITY");
-                    }
-                    if (element.element("REGION") != null) {
-                        homeState = element.elementTextTrim("REGION");
-                    }
-                    if (element.element("PCODE") != null) {
-                        homeZip = element.elementTextTrim("PCODE");
-                    }
-                    if (element.element("CTRY") != null) {
-                        homeCountry = element.elementTextTrim("CTRY");
-                    }
-                }
-                else if (element.element("WORK") != null) {
-                    if (element.element("STREET") != null) {
-                        businessStreet = element.elementTextTrim("STREET");
-                    }
-                    if (element.element("LOCALITY") != null) {
-                        businessCity = element.elementTextTrim("LOCALITY");
-                    }
-                    if (element.element("REGION") != null) {
-                        businessState = element.elementTextTrim("REGION");
-                    }
-                    if (element.element("PCODE") != null) {
-                        businessZip = element.elementTextTrim("PCODE");
-                    }
-                    if (element.element("CTRY") != null) {
-                        businessCountry = element.elementTextTrim("CTRY");
-                    }
-                }
-            }
-            // Parse telephones
-            Iterator<Element> telephones = vCard.elementIterator("TEL");
-            while (telephones.hasNext()) {
-                element = telephones.next();
-                if (element.element("HOME") != null) {
-                    if (element.element("VOICE") != null) {
-                        homePhone = element.elementTextTrim("NUMBER");
-                    }
-                    else if (element.element("CELL") != null) {
-                        homeMobile = element.elementTextTrim("NUMBER");
-                    }
-                    else if (element.element("FAX") != null) {
-                        homeFax = element.elementTextTrim("NUMBER");
-                    }
-                    else if (element.element("PAGER") != null) {
-                        homePager = element.elementTextTrim("NUMBER");
-                    }
-                }
-                else if (element.element("WORK") != null) {
-                    if (element.element("VOICE") != null) {
-                        businessPhone = element.elementTextTrim("NUMBER");
-                    }
-                    else if (element.element("CELL") != null) {
-                        businessMobile = element.elementTextTrim("NUMBER");
-                    }
-                    else if (element.element("FAX") != null) {
-                        businessFax = element.elementTextTrim("NUMBER");
-                    }
-                    else if (element.element("PAGER") != null) {
-                        businessPager = element.elementTextTrim("NUMBER");
-                    }
-                }
-            }
-            element = vCard.element("TITLE");
-            if (element != null) {
-                businessJobTitle = element.getTextTrim();
-            }
-            element = vCard.element("ORG");
-            if (element != null) {
-                if (element.element("ORGUNIT") != null) {
-                    businessDepartment = element.elementTextTrim("ORGUNIT");
-                }
-            }
+            loadFromElement(vCard);
             avatarStoredInDB = LdapVCardProvider.STORE_AVATAR_IN_DB.getValue();
         }
         catch (DocumentException e) {
