+++ xmppserver/src/main/java/org/jivesoftware/openfire/vcard/VCardBean.java
@@ -1,0 +1,524 @@
+/*
+ * Copyright (C) 2005-2008 Jive Software, 2017-2025 Ignite Realtime Foundation. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jivesoftware.openfire.vcard;
+
+import org.dom4j.DocumentHelper;
+import org.dom4j.Element;
+import org.dom4j.QName;
+
+import java.util.Iterator;
+
+public class VCardBean
+{
+    protected String name = ""; // NOTE: Openfire traditionally uses the vCard GIVEN name field as its 'name' field (and appears to disregard FAMILY and MIDDLE)
+    protected String email = "";
+    protected String fullName = "";
+    protected String nickname = "";
+    protected String birthday = "";
+    protected String photo = "";
+    protected String photoType = "";
+    protected String homeStreet = "";
+    protected String homeCity = "";
+    protected String homeState = "";
+    protected String homeZip = "";
+    protected String homeCountry = "";
+    protected String homePhone = "";
+    protected String homeMobile = "";
+    protected String homeFax = "";
+    protected String homePager = "";
+    protected String businessStreet = "";
+    protected String businessCity = "";
+    protected String businessState = "";
+    protected String businessZip = "";
+    protected String businessCountry = "";
+    protected String businessJobTitle = "";
+    protected String businessDepartment = "";
+    protected String businessPhone = "";
+    protected String businessMobile = "";
+    protected String businessFax = "";
+    protected String businessPager = "";
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getFullName() {
+        return fullName;
+    }
+
+    public void setFullName(String fullName) {
+        this.fullName = fullName;
+    }
+
+    public String getNickname() {
+        return nickname;
+    }
+
+    public void setNickname(String nickname) {
+        this.nickname = nickname;
+    }
+
+    public String getBirthday() {
+        return birthday;
+    }
+
+    public void setBirthday(String birthday) {
+        this.birthday = birthday;
+    }
+
+    public String getPhoto() {
+        return photo;
+    }
+
+    public void setPhoto(String photo) {
+        this.photo = photo;
+    }
+
+    public String getPhotoType() {
+        return photoType;
+    }
+
+    public void setPhotoType(String photoType) {
+        this.photoType = photoType;
+    }
+
+    public String getHomeStreet() {
+        return homeStreet;
+    }
+
+    public void setHomeStreet(String homeStreet) {
+        this.homeStreet = homeStreet;
+    }
+
+    public String getHomeCity() {
+        return homeCity;
+    }
+
+    public void setHomeCity(String homeCity) {
+        this.homeCity = homeCity;
+    }
+
+    public String getHomeState() {
+        return homeState;
+    }
+
+    public void setHomeState(String homeState) {
+        this.homeState = homeState;
+    }
+
+    public String getHomeZip() {
+        return homeZip;
+    }
+
+    public void setHomeZip(String homeZip) {
+        this.homeZip = homeZip;
+    }
+
+    public String getHomeCountry() {
+        return homeCountry;
+    }
+
+    public void setHomeCountry(String homeCountry) {
+        this.homeCountry = homeCountry;
+    }
+
+    public String getHomePhone() {
+        return homePhone;
+    }
+
+    public void setHomePhone(String homePhone) {
+        this.homePhone = homePhone;
+    }
+
+    public String getHomeMobile() {
+        return homeMobile;
+    }
+
+    public void setHomeMobile(String homeMobile) {
+        this.homeMobile = homeMobile;
+    }
+
+    public String getHomeFax() {
+        return homeFax;
+    }
+
+    public void setHomeFax(String homeFax) {
+        this.homeFax = homeFax;
+    }
+
+    public String getHomePager() {
+        return homePager;
+    }
+
+    public void setHomePager(String homePager) {
+        this.homePager = homePager;
+    }
+
+    public String getBusinessStreet() {
+        return businessStreet;
+    }
+
+    public void setBusinessStreet(String businessStreet) {
+        this.businessStreet = businessStreet;
+    }
+
+    public String getBusinessCity() {
+        return businessCity;
+    }
+
+    public void setBusinessCity(String businessCity) {
+        this.businessCity = businessCity;
+    }
+
+    public String getBusinessState() {
+        return businessState;
+    }
+
+    public void setBusinessState(String businessState) {
+        this.businessState = businessState;
+    }
+
+    public String getBusinessZip() {
+        return businessZip;
+    }
+
+    public void setBusinessZip(String businessZip) {
+        this.businessZip = businessZip;
+    }
+
+    public String getBusinessCountry() {
+        return businessCountry;
+    }
+
+    public void setBusinessCountry(String businessCountry) {
+        this.businessCountry = businessCountry;
+    }
+
+    public String getBusinessJobTitle() {
+        return businessJobTitle;
+    }
+
+    public void setBusinessJobTitle(String businessJobTitle) {
+        this.businessJobTitle = businessJobTitle;
+    }
+
+    public String getBusinessDepartment() {
+        return businessDepartment;
+    }
+
+    public void setBusinessDepartment(String businessDepartment) {
+        this.businessDepartment = businessDepartment;
+    }
+
+    public String getBusinessPhone() {
+        return businessPhone;
+    }
+
+    public void setBusinessPhone(String businessPhone) {
+        this.businessPhone = businessPhone;
+    }
+
+    public String getBusinessMobile() {
+        return businessMobile;
+    }
+
+    public void setBusinessMobile(String businessMobile) {
+        this.businessMobile = businessMobile;
+    }
+
+    public String getBusinessFax() {
+        return businessFax;
+    }
+
+    public void setBusinessFax(String businessFax) {
+        this.businessFax = businessFax;
+    }
+
+    public String getBusinessPager() {
+        return businessPager;
+    }
+
+    public void setBusinessPager(String businessPager) {
+        this.businessPager = businessPager;
+    }
+
+    public Element asElement() {
+        Element vCard = DocumentHelper.createElement(QName.get("vCard", "vcard-temp"));
+        Element subelement;
+
+        // Add name
+        if (name != null && !name.trim().isEmpty()) {
+            subelement = vCard.addElement("N");
+            subelement.addElement("GIVEN").setText(name.trim());
+        }
+        // Add email
+        if (email != null && !email.trim().isEmpty()) {
+            subelement = vCard.addElement("EMAIL");
+            subelement.addElement("INTERNET");
+            subelement.addElement("USERID").setText(email.trim());
+        }
+        // Add Full Name
+        vCard.addElement("FN").setText(fullName.trim());
+        // Add nickname
+        if (nickname != null && !nickname.trim().isEmpty()) {
+            vCard.addElement("NICKNAME").setText(nickname.trim());
+        }
+        // Add birthday
+        if (birthday != null && !birthday.trim().isEmpty()) {
+            vCard.addElement("BDAY").setText(birthday.trim());
+        }
+        // Add photo/avatar
+        if (photo != null && !photo.trim().isEmpty()) {
+            Element element = vCard.addElement("PHOTO");
+            element.addElement("TYPE").setText(photoType == null || photoType.trim().isEmpty() ?  "image/jpeg" : photoType.trim());
+            element.addElement("BINVAL").setText(photo.trim());
+        }
+        // Add home address
+        subelement = vCard.addElement("ADR");
+        subelement.addElement("HOME");
+        if (homeStreet != null && !homeStreet.trim().isEmpty()) {
+            subelement.addElement("STREET").setText(homeStreet.trim());
+        }
+        if (homeCity != null && !homeCity.trim().isEmpty()) {
+            subelement.addElement("LOCALITY").setText(homeCity.trim());
+        }
+        if (homeState != null && !homeState.trim().isEmpty()) {
+            subelement.addElement("REGION").setText(homeState.trim());
+        }
+        if (homeZip != null && !homeZip.trim().isEmpty()) {
+            subelement.addElement("PCODE").setText(homeZip.trim());
+        }
+        if (homeCountry != null && !homeCountry.trim().isEmpty()) {
+            subelement.addElement("CTRY").setText(homeCountry.trim());
+        }
+        // Add business address
+        subelement = vCard.addElement("ADR");
+        subelement.addElement("WORK");
+        if (businessStreet != null && !businessStreet.trim().isEmpty()) {
+            subelement.addElement("STREET").setText(businessStreet.trim());
+        }
+        if (businessCity != null && !businessCity.trim().isEmpty()) {
+            subelement.addElement("LOCALITY").setText(businessCity.trim());
+        }
+        if (businessState != null && !businessState.trim().isEmpty()) {
+            subelement.addElement("REGION").setText(businessState.trim());
+        }
+        if (businessZip != null && !businessZip.trim().isEmpty()) {
+            subelement.addElement("PCODE").setText(businessZip.trim());
+        }
+        if (businessCountry != null && !businessCountry.trim().isEmpty()) {
+            subelement.addElement("CTRY").setText(businessCountry.trim());
+        }
+        // Add home phone
+        if (homePhone != null && !homePhone.trim().isEmpty()) {
+            subelement = vCard.addElement("TEL");
+            subelement.addElement("HOME");
+            subelement.addElement("VOICE");
+            subelement.addElement("NUMBER").setText(homePhone.trim());
+        }
+        // Add home mobile
+        if (homeMobile != null && !homeMobile.trim().isEmpty()) {
+            subelement = vCard.addElement("TEL");
+            subelement.addElement("HOME");
+            subelement.addElement("CELL");
+            subelement.addElement("NUMBER").setText(homeMobile.trim());
+        }
+        // Add home fax
+        if (homeFax != null && !homeFax.trim().isEmpty()) {
+            subelement = vCard.addElement("TEL");
+            subelement.addElement("HOME");
+            subelement.addElement("FAX");
+            subelement.addElement("NUMBER").setText(homeFax.trim());
+        }
+        // Add home pager
+        if (homePager != null && !homePager.trim().isEmpty()) {
+            subelement = vCard.addElement("TEL");
+            subelement.addElement("HOME");
+            subelement.addElement("PAGER");
+            subelement.addElement("NUMBER").setText(homePager.trim());
+        }
+        // Add business phone
+        if (businessPhone != null && !businessPhone.trim().isEmpty()) {
+            subelement = vCard.addElement("TEL");
+            subelement.addElement("WORK");
+            subelement.addElement("VOICE");
+            subelement.addElement("NUMBER").setText(businessPhone.trim());
+        }
+        // Add business mobile
+        if (businessMobile != null && !businessMobile.trim().isEmpty()) {
+            subelement = vCard.addElement("TEL");
+            subelement.addElement("WORK");
+            subelement.addElement("CELL");
+            subelement.addElement("NUMBER").setText(businessMobile.trim());
+        }
+        // Add business fax
+        if (businessFax != null && !businessFax.trim().isEmpty()) {
+            subelement = vCard.addElement("TEL");
+            subelement.addElement("WORK");
+            subelement.addElement("FAX");
+            subelement.addElement("NUMBER").setText(businessFax.trim());
+        }
+        // Add business pager
+        if (businessPager != null && !businessPager.trim().isEmpty()) {
+            subelement = vCard.addElement("TEL");
+            subelement.addElement("WORK");
+            subelement.addElement("PAGER");
+            subelement.addElement("NUMBER").setText(businessPager.trim());
+        }
+        // Add job title
+        if (businessJobTitle != null && !businessJobTitle.trim().isEmpty()) {
+            vCard.addElement("TITLE").setText(businessJobTitle.trim());
+        }
+        // Add job department
+        if (businessDepartment != null && !businessDepartment.trim().isEmpty()) {
+            vCard.addElement("ORG").addElement("ORGUNIT").setText(businessDepartment.trim());
+        }
+
+        return vCard;
+    }
+
+    public void loadFromElement(final Element vCard) {
+        if (vCard == null) {
+            return;
+        }
+        Element element = vCard.element("N");
+        if (element != null) {
+            name = element.elementTextTrim("GIVEN");
+        }
+        element = vCard.element("EMAIL");
+        if (element != null) {
+            email = element.elementTextTrim("USERID");
+        }
+        element = vCard.element("FN");
+        if (element != null) {
+            fullName = element.getTextTrim();
+        }
+        element = vCard.element("NICKNAME");
+        if (element != null) {
+            nickname = element.getTextTrim();
+        }
+        element = vCard.element("BDAY");
+        if (element != null) {
+            birthday = element.getTextTrim();
+        }
+        element = vCard.element("PHOTO");
+        if (element != null) {
+            element = vCard.element("PHOTO").element("BINVAL");
+            if (element != null) {
+                photo = element.getTextTrim();
+            }
+            element = vCard.element("PHOTO").element("TYPE");
+            if (element != null) {
+                photoType = element.getTextTrim();
+            }
+        }
+
+        // Parse addresses
+        Iterator<Element> addresses = vCard.elementIterator("ADR");
+        while (addresses.hasNext()) {
+            element = addresses.next();
+            if (element.element("HOME") != null) {
+                if (element.element("STREET") != null) {
+                    homeStreet = element.elementTextTrim("STREET");
+                }
+                if (element.element("LOCALITY") != null) {
+                    homeCity = element.elementTextTrim("LOCALITY");
+                }
+                if (element.element("REGION") != null) {
+                    homeState = element.elementTextTrim("REGION");
+                }
+                if (element.element("PCODE") != null) {
+                    homeZip = element.elementTextTrim("PCODE");
+                }
+                if (element.element("CTRY") != null) {
+                    homeCountry = element.elementTextTrim("CTRY");
+                }
+            }
+            else if (element.element("WORK") != null) {
+                if (element.element("STREET") != null) {
+                    businessStreet = element.elementTextTrim("STREET");
+                }
+                if (element.element("LOCALITY") != null) {
+                    businessCity = element.elementTextTrim("LOCALITY");
+                }
+                if (element.element("REGION") != null) {
+                    businessState = element.elementTextTrim("REGION");
+                }
+                if (element.element("PCODE") != null) {
+                    businessZip = element.elementTextTrim("PCODE");
+                }
+                if (element.element("CTRY") != null) {
+                    businessCountry = element.elementTextTrim("CTRY");
+                }
+            }
+        }
+        // Parse telephones
+        Iterator<Element> telephones = vCard.elementIterator("TEL");
+        while (telephones.hasNext()) {
+            element = telephones.next();
+            if (element.element("HOME") != null) {
+                if (element.element("VOICE") != null) {
+                    homePhone = element.elementTextTrim("NUMBER");
+                }
+                else if (element.element("CELL") != null) {
+                    homeMobile = element.elementTextTrim("NUMBER");
+                }
+                else if (element.element("FAX") != null) {
+                    homeFax = element.elementTextTrim("NUMBER");
+                }
+                else if (element.element("PAGER") != null) {
+                    homePager = element.elementTextTrim("NUMBER");
+                }
+            }
+            else if (element.element("WORK") != null) {
+                if (element.element("VOICE") != null) {
+                    businessPhone = element.elementTextTrim("NUMBER");
+                }
+                else if (element.element("CELL") != null) {
+                    businessMobile = element.elementTextTrim("NUMBER");
+                }
+                else if (element.element("FAX") != null) {
+                    businessFax = element.elementTextTrim("NUMBER");
+                }
+                else if (element.element("PAGER") != null) {
+                    businessPager = element.elementTextTrim("NUMBER");
+                }
+            }
+        }
+        element = vCard.element("TITLE");
+        if (element != null) {
+            businessJobTitle = element.getTextTrim();
+        }
+        element = vCard.element("ORG");
+        if (element != null) {
+            if (element.element("ORGUNIT") != null) {
+                businessDepartment = element.elementTextTrim("ORGUNIT");
+            }
+        }
+    }
+}
